0023934: Compiler warnings in MS VC++ 10
Elimination of compiler warnings
- ExprIntrp: generated file ExprIntrp.tab.c and related WOK scripts removed from the sources
- ExprIntrp, StepFile: added missing declarations and casts; warnings caused by Flex and Bison code suppressed for MSVC compuler by #pragma
- OSD: dummy #includes added to files containing no code for Windows, to avoid warning on empty file
- PLib: piece of code contained in PLib_ChangeDim.gxx included explicitly in cxx and cleaned
- Other places: some casts added to avoid warnings

+++ src/Aspect/Aspect.cxx
@@ -113,7 +113,7 @@ Standard_PCharacter pstring;
          estring.SetValue(i,c);
       } else if( (c > Standard_ExtCharacter(0x0020)) && 
 			(c <= Standard_ExtCharacter(0x007F)) ) {
-	c += Standard_ExtCharacter(0xFEE0);
+	c += Standard_ExtCharacter((unsigned short)0xFEE0);
 	estring.SetValue(i,c);
       }
     }
